Improve player prefab
Makes the player hitbox better centered and more accurate so the player can no longer clip into walls, and walking against a wall to the left looks the same as walking against a wall to the right.

Makes the player-tracking camera into part of the prefab, so we no longer have to re-link it after importing into new scenes. We can do this now that we no longer have per-scene camera requirements such as the shark dock.

Changes camera angle, distance and offset to increase visibility and viewing comfort.