Add player animation switching
The player prefab now switches between eight animations for walking in the four cardinal directions and standing after walking in the four cardinal directions. When moving diagonally, direction is determined by whichever axis the movement is closest to, with east/west used over north/south as a tiebreaker. I don't know if this is the "proper" way to do this (having multiple sprite renderers, and enabling/disabling them as they are needed) but it works. Stand North and Stand South animations use other animations as placeholders because I couldn't find the real ones.