Implement the FOHI
The FOHI is now in-game, animated, and follows you around. It's not very smart and can easily get stuck, so be nice to the FOHI.

When importing the prefab to scenes, make sure to set the Tracked Player property to whoever you want the FOHI to follow. Usually the player, but it could follow a tree instead if you want.