Add spawn point to scene transitions
Scene transitions have a field to pass in spawn point data for player and fohi now. This needs to be set or else the player will spawn at 0,0,0. Currently only has data set for the transition between southeast coast and player camp, will set the rest in a followup PR. Also flattens southeast coast to remove the ramp.