Dirty fix for rotated player controls
The player controls were messed up in the map dev scene because the entire map (and thus camera) are rotated 90 degrees compared to other scenes. This fix applies a simple transform to the player's input to make them appear to move in the right direction again. However the scene is still rotated, please discuss whether this should be "properly" fixed before closing the associated issue.