Fix some player animation stuff
Connects the sprite renderer in map dev for the player the way it should be according to the prefab, and corrects an error in the direction detection logic in player controller so that it uses the player's input x and y values instead of the rotated ones. Player animations in map dev appear to work smoothly now.