Add parameters to interactable script
Add parameters for the interactable script so it can be activated by proximity, proximity + interact button, or mouse click, and so it can do any combination of the following: run a yarn node, teleport the player, and give the player an item (not yet implemented). Add triggers over worm dirt that give items with tag Worm. Add a teleport to the first customer to move the player to the south dock after speaking.